path-routing-algorithm/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.0
pydantic==2.4.2
python-dotenv==1.0.0
loguru==0.7.2